pyproject.toml
[tool.poetry]
name = "sonrisas-backend"
version = "0.1.0"
description = ""
authors = ["Solano Jason <pajason2000@outlook.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.115.12"}
alembic = "^1.15.2"
pydantic = "^2.11.2"
pydantic-settings = "^2.8.1"
psycopg = {extras = ["binary"], version = "^3.2.6"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.40"}
google-cloud-storage = "^3.1.0"
sqlalchemy-file = "^0.6.0"
fasteners = "^0.19"
apache-libcloud = "^3.8.0"
cryptography = "^44.0.2"
pillow = "^11.1.0"
black = "^25.1.0"
pytest = "^8.3.5"
sqlalchemy-utils = "^0.41.2"


[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.4.0"
qrcode = {extras = ["pil"], version = "^8.1"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
